moved classes to better packages
